+++ src/components/AddWaterBtn/AddWaterBtn.module.css
@@ -1,23 +1,3 @@
-.wrapper {
-  background-color: var(--background-color);
-  color: var(--main-text);
-  padding: 1.25rem;
-  margin-bottom: 2.5rem;
-}
-
-.cardHeader {
-  display: flex;
-  justify-content: space-between;
-  margin-bottom: 1.5rem;
-}
-
-.cardTitle {
-  font-weight: 700;
-  font-size: 1.25rem;
-  line-height: 120%;
-  letter-spacing: -0.01em;
-}
-
 .cardButton {
   display: flex;
   justify-content: flex-start;
@@ -57,19 +37,4 @@
   line-height: 150%;
   letter-spacing: -0.01em;
   text-align: center;
-}
-
-@media screen and (min-width: 768px) {
-  .wrapper {
-    padding: 2rem;
-  }
-  .cardTitle {
-    font-size: 1.75rem;
-    line-height: 114%;
-  }
-}
-@media screen and (min-width: 1440px) {
-  .wrapper {
-    margin-bottom: 4.5rem;
-  }
 }
