+++ src/components/ChooseDate/ChooseDate.module.css
@@ -1,16 +1,3 @@
-.wrapper {
-  background-color: var(--background-color);
-  color: var(--main-text);
-  padding: 1.25rem;
-  margin-bottom: 2.5rem;
-}
-
-.cardHeader {
-  display: flex;
-  justify-content: space-between;
-  margin-bottom: 1.5rem;
-}
-
 .cardTitle {
   font-weight: 700;
   font-size: 1.25rem;
@@ -18,58 +5,9 @@
   letter-spacing: -0.01em;
 }
 
-.cardButton {
-  display: flex;
-  justify-content: flex-start;
-  align-items: center;
-  column-gap: 0.625rem;
-  background-color: inherit;
-  color: inherit;
-  font-size: inherit;
-  font-weight: inherit;
-  padding: 0px;
-  border: 0px;
-}
-
-.cardIcon {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 30px;
-  height: 30px;
-  background-color: var(--accent-color);
-  border-radius: 50%;
-  stroke: var(--main-text);
-  transition-property: background-color, stroke;
-  transition-duration: 250ms;
-  transition-timing-function: ease;
-}
-
-.cardButton:hover .cardIcon,
-.cardButton:focus .cardIcon {
-  background-color: var(--main-text);
-  stroke: var(--main-white);
-}
-
-.cardButtonTitle {
-  font-weight: 700;
-  font-size: 1rem;
-  line-height: 150%;
-  letter-spacing: -0.01em;
-  text-align: center;
-}
-
 @media screen and (min-width: 768px) {
-  .wrapper {
-    padding: 2rem;
-  }
   .cardTitle {
     font-size: 1.75rem;
     line-height: 114%;
-  }
-}
-@media screen and (min-width: 1440px) {
-  .wrapper {
-    margin-bottom: 4.5rem;
   }
 }
